LayoutTests/svg/dynamic-updates/SVGUseElement-dom-requiredFeatures-expected.txt

SVG 1.1 dynamic update tests

Check that setting requiredFeatures to something valid makes it render again

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.defaultView.getComputedStyle(useElement, null).display is "inline"
PASS document.defaultView.getComputedStyle(useElement, null).display is ""
PASS document.defaultView.getComputedStyle(useElement, null).display is "inline"
Check that adding something valid to requiredFeatures keeps rendering the element
PASS document.defaultView.getComputedStyle(useElement, null).display is "inline"
Check that adding something invalid to requiredFeatures makes it not render
PASS document.defaultView.getComputedStyle(useElement, null).display is ""
PASS successfullyParsed is true

TEST COMPLETE
